dispatch2

Struct DispatchData

pub struct DispatchData { /* private fields */ }

Dispatch data.

Implementations

impl DispatchData

pub fn empty() -> &'static Self

Get an empty DispatchData.

impl DispatchData

pub fn size(&self) -> usize

Returns the logical size of the memory region(s) represented by the specified dispatch data object.

Parameter data: The dispatch data object to query.

Returns: The number of bytes represented by the data object.

pub unsafe fn map( &self, buffer_ptr: *mut *const c_void, size_ptr: *mut usize, ) -> DispatchRetained<DispatchData>

Maps the memory represented by the specified dispatch data object as a single contiguous memory region and returns a new data object representing it. If non-NULL references to a pointer and a size variable are provided, they are filled with the location and extent of that region. These allow direct read access to the represented memory, but are only valid until the returned object is released. Under ARC, if that object is held in a variable with automatic storage, care needs to be taken to ensure that it is not released by the compiler before memory access via the pointer has been completed.

Parameter data: The dispatch data object to map.

Parameter buffer_ptr: A pointer to a pointer variable to be filled with the location of the mapped contiguous memory region, or NULL.

Parameter size_ptr: A pointer to a size_t variable to be filled with the size of the mapped contiguous memory region, or NULL.

Returns: A newly created dispatch data object.

Safety

buffer_ptr must be a valid pointer or null.

size_ptr must be a valid pointer or null.

pub fn concat(&self, data2: &DispatchData) -> DispatchRetained<DispatchData>

Returns a new dispatch data object representing the concatenation of the specified data objects. Those objects may be released by the application after the call returns (however, the system might not deallocate the memory region(s) described by them until the newly created object has also been released).

Parameter data1: The data object representing the region(s) of memory to place at the beginning of the newly created object.

Parameter data2: The data object representing the region(s) of memory to place at the end of the newly created object.

Returns: A newly created object representing the concatenation of the data1 and data2 objects.

pub unsafe fn subrange( &self, offset: usize, length: usize, ) -> DispatchRetained<DispatchData>

Returns a new dispatch data object representing a subrange of the specified data object, which may be released by the application after the call returns (however, the system might not deallocate the memory region(s) described by that object until the newly created object has also been released).

Parameter data: The data object representing the region(s) of memory to create a subrange of.

Parameter offset: The offset into the data object where the subrange starts.

Parameter length: The length of the range.

Returns: A newly created object representing the specified subrange of the data object.

impl DispatchData

pub unsafe fn region( &self, location: usize, offset_ptr: NonNull<usize>, ) -> DispatchRetained<DispatchData>

Finds the contiguous memory region containing the specified location among the regions represented by the specified object and returns a copy of the internal dispatch data object representing that region along with its logical offset in the specified object.

Parameter data: The dispatch data object to query.

Parameter location: The logical position in the data object to query.

Parameter offset_ptr: A pointer to a size_t variable to be filled with the logical offset of the returned region object to the start of the queried data object.

Returns: A newly created dispatch data object.

Safety

offset_ptr must be a valid pointer.