dispatch2

Struct DispatchIO

pub struct DispatchIO { /* private fields */ }

Dispatch IO.

Implementations

impl DispatchIO

pub fn close(&self, flags: DispatchIOCloseFlags)

Close the specified I/O channel to new read or write operations; scheduling operations on a closed channel results in their handler returning an error.

If the DISPATCH_IO_STOP flag is provided, the system will make a best effort to interrupt any outstanding read and write operations on the I/O channel, otherwise those operations will run to completion normally. Partial results of read and write operations may be returned even after a channel is closed with the DISPATCH_IO_STOP flag. The final invocation of an I/O handler of an interrupted operation will be passed an ECANCELED error code, as will the I/O handler of an operation scheduled on a closed channel.

Parameter channel: The dispatch I/O channel to close.

Parameter flags: The flags for the close operation.

pub fn descriptor(&self) -> dispatch_fd_t

Returns the file descriptor underlying a dispatch I/O channel.

Will return -1 for a channel closed with dispatch_io_close() and for a channel associated with a path name that has not yet been open(2)ed.

If called from a barrier block scheduled on a channel associated with a path name that has not yet been open(2)ed, this will trigger the channel open(2) operation and return the resulting file descriptor.

Parameter channel: The dispatch I/O channel to query.

Returns: The file descriptor underlying the channel, or -1.

pub fn set_high_water(&self, high_water: usize)

Set a high water mark on the I/O channel for all operations.

The system will make a best effort to enqueue I/O handlers with partial results as soon the number of bytes processed by an operation (i.e. read or written) reaches the high water mark.

The size of data objects passed to I/O handlers for this channel will never exceed the specified high water mark.

The default value for the high water mark is unlimited (i.e. SIZE_MAX).

Parameter channel: The dispatch I/O channel on which to set the policy.

Parameter high_water: The number of bytes to use as a high water mark.

pub fn set_low_water(&self, low_water: usize)

Set a low water mark on the I/O channel for all operations.

The system will process (i.e. read or write) at least the low water mark number of bytes for an operation before enqueueing I/O handlers with partial results.

The size of data objects passed to intermediate I/O handler invocations for this channel (i.e. excluding the final invocation) will never be smaller than the specified low water mark, except if the channel has an interval with the DISPATCH_IO_STRICT_INTERVAL flag set or if EOF or an error was encountered.

I/O handlers should be prepared to receive amounts of data significantly larger than the low water mark in general. If an I/O handler requires intermediate results of fixed size, set both the low and and the high water mark to that size.

The default value for the low water mark is unspecified, but must be assumed to be such that intermediate handler invocations may occur. If I/O handler invocations with partial results are not desired, set the low water mark to SIZE_MAX.

Parameter channel: The dispatch I/O channel on which to set the policy.

Parameter low_water: The number of bytes to use as a low water mark.

pub fn set_interval(&self, interval: u64, flags: DispatchIOIntervalFlags)

Set a nanosecond interval at which I/O handlers are to be enqueued on the I/O channel for all operations.

This allows an application to receive periodic feedback on the progress of read and write operations, e.g. for the purposes of displaying progress bars.

If the amount of data ready to be delivered to an I/O handler at the interval is inferior to the channel low water mark, the handler will only be enqueued if the DISPATCH_IO_STRICT_INTERVAL flag is set.

Note that the system may defer enqueueing interval I/O handlers by a small unspecified amount of leeway in order to align with other system activity for improved system performance or power consumption.

Parameter channel: The dispatch I/O channel on which to set the policy.

Parameter interval: The interval in nanoseconds at which delivery of the I/O handler is desired.

Parameter flags: Flags indicating desired data delivery behavior at interval time.