dispatch2

Struct DispatchRetained

pub struct DispatchRetained<T: ?Sized> { /* private fields */ }

A reference counted pointer type for Dispatch objects.

DispatchRetained strongly references or โretainsโ the given object T, and decrements the retain count or โreleasesโ it again when dropped, thereby ensuring it will be deallocated at the right time.

The type T inside DispatchRetained<T> can be anything that implements DispatchObject, i.e. any Dispatch object.

Comparison to other types

DispatchRetained<T> is equivalent to objc2::rc::Retained, and can be converted to/from that when the "objc2" feature is enabled. Note though that this type uses the underlying Dispatch primitives dispatch_retain / dispatch_release instead of objc_retain / objc_release for performance, and to avoid depending on the Objective-C runtime if not needed.

You can also view DispatchRetained<T> as the Dispatch equivalent of std::sync::Arc, that is, it is a thread-safe reference-counting smart pointer that allows cloning by bumping the reference count.

Unlike Arc, objects can be retained directly from a &T using DispatchObject::retain (for Arc you need &Arc<T>).

Weak references are not supported, for that you need to convert to objc2::rc::Retained.

Forwarding implementations

Since DispatchRetained<T> is a smart pointer, it Derefs to T.

It also forwards the implementation of a bunch of standard library traits such as PartialEq, AsRef, and so on, so that it becomes possible to use e.g. DispatchRetained<DispatchQueue> as-if it you had a &DispatchQueue. Note that having a DispatchQueue directly is not possible since dispatch objects cannot live on the stack, but instead must reside on the heap, and as such must be accessed behind a pointer or a reference.

Memory layout

This is guaranteed to have the same size and alignment as a pointer to the object, i.e. same as *const T or dispatch_object_t.

Additionally, it participates in the null-pointer optimization, that is, Option<DispatchRetained<T>> is guaranteed to have the same size as DispatchRetained<T>.

Implementations

impl<T: ?Sized + DispatchObject> DispatchRetained<T>

pub unsafe fn from_raw(ptr: NonNull<T>) -> Self

Construct a DispatchRetained from a pointer that already has +1 retain count.

This is useful when you have been given a pointer to an object from some API that returns a retained pointer, and expects you to release it. That is, an API that follows the create rule.

Safety

The pointer must be a valid and live object according to Dispatch, and it must be of type T.

Additionally, you must ensure the given object pointer has +1 retain count.

pub unsafe fn retain(ptr: NonNull<T>) -> Self

Retain the pointer and construct a DispatchRetained from it.

This is useful when you have been given a pointer to an object from some API that follows the get rule, and you would like to keep it around for longer than the current memory context.

Safety

The pointer must be a valid and live object according to Dispatch, and it must be of type T.

pub fn into_raw(this: Self) -> NonNull<T>

Consumes the DispatchRetained, returning a raw pointer with +1 retain count.

After calling this function, the caller is responsible for the memory previously managed by the DispatchRetained.

This is effectively the opposite of DispatchRetained::from_raw.

This is an associated method, and must be called as DispatchRetained::into_raw(obj).

pub fn as_ptr(this: &Self) -> NonNull<T>

Returns a raw pointer to the object.

The pointer is valid for at least as long as the DispatchRetained is held.

This is an associated method, and must be called as DispatchRetained::as_ptr(&obj).

pub unsafe fn cast_unchecked<U: DispatchObject>( this: Self, ) -> DispatchRetained<U>

Unchecked conversion to another Dispatch object.

This is equivalent to a cast between two pointers.

This is an associated method, and must be called as DispatchRetained::cast_unchecked(obj).

Safety

You must ensure that the object can be reinterpreted as the given type.

If T is not 'static, you must ensure that U ensures that the data contained by T is kept alive for as long as U lives.

Additionally, you must ensure that any safety invariants that the new type has are upheld.