dispatch2

Struct DispatchSemaphore

pub struct DispatchSemaphore { /* private fields */ }

Dispatch semaphore.

Implementations

impl DispatchSemaphore

pub fn new(value: isize) -> DispatchRetained<DispatchSemaphore>

Creates new counting semaphore with an initial value.

Passing zero for the value is useful for when two threads need to reconcile the completion of a particular event. Passing a value greater than zero is useful for managing a finite pool of resources, where the pool size is equal to the value.

Parameter value: The starting value for the semaphore. Passing a value less than zero will cause NULL to be returned.

Returns: The newly created semaphore, or NULL on failure.

pub fn wait(&self, timeout: DispatchTime) -> isize

Wait (decrement) for a semaphore.

Decrement the counting semaphore. If the resulting value is less than zero, this function waits for a signal to occur before returning. If the timeout is reached without a signal being received, the semaphore is re-incremented before the function returns.

Parameter dsema: The semaphore. The result of passing NULL in this parameter is undefined.

Parameter timeout: When to timeout (see dispatch_time). As a convenience, there are the DISPATCH_TIME_NOW and DISPATCH_TIME_FOREVER constants.

Returns: Returns zero on success, or non-zero if the timeout occurred.

pub fn signal(&self) -> isize

Signal (increment) a semaphore.

Increment the counting semaphore. If the previous value was less than zero, this function wakes a waiting thread before returning.

Parameter dsema: The counting semaphore. The result of passing NULL in this parameter is undefined.

Returns: This function returns non-zero if a thread is woken. Otherwise, zero is returned.

impl DispatchSemaphore

pub fn try_acquire( &self, timeout: DispatchTime, ) -> Result<DispatchSemaphoreGuard, WaitError>

Attempt to acquire the DispatchSemaphore and return a DispatchSemaphoreGuard.

Errors

Return WaitError::TimeOverflow if the passed timeout is too big.

Return WaitError::Timeout in case of timeout.