dispatch2

Struct DispatchWorkloop

pub struct DispatchWorkloop { /* private fields */ }

Dispatch workloop queue.

Implementations

impl DispatchWorkloop

pub fn set_autorelease_frequency(&self, frequency: DispatchAutoReleaseFrequency)

Sets the autorelease frequency of the workloop.

See dispatch_queue_attr_make_with_autorelease_frequency(). The default policy for a workloop is DISPATCH_AUTORELEASE_FREQUENCY_WORK_ITEM.

Parameter workloop: The dispatch workloop to modify.

This workloop must be inactive, passing an activated object is undefined and will cause the process to be terminated.

Parameter frequency: The requested autorelease frequency.

impl DispatchWorkloop

pub fn new(label: &str, inactive: bool) -> DispatchRetained<Self>

Create a new DispatchWorkloop.